[Fix] Gender자료형 수정

+++ src/main/java/com/potato/balbambalbam/main/cardInfo/dto/VoiceRequestDto.java
@@ -8,9 +8,9 @@ import lombok.ToString;
 @ToString
 public class VoiceRequestDto {
     public Integer age;
-    public Integer gender;  //0 : 여자 , 1 : 남자
+    public Byte gender;  //0 : 여자 , 1 : 남자
 
-    public VoiceRequestDto(Integer age, Integer gender) {
+    public VoiceRequestDto(Integer age, Byte gender) {
         this.age = age;
         this.gender = gender;
     }
